# easytravel /api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  models.Country:
    properties:
      code:
        type: string
      continent:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.CountryPrimaryKey:
    properties:
      id:
        type: string
    type: object
...
  models.CreateCountry:
    properties:
      code:
        type: string
      continent:
        type: string
      title:
        type: string
    type: object
...
  models.GetListCounrtyResponse:
    properties:
      count:
        type: integer
      countries:
        items:
          $ref: '#/definitions/models.Country'
        type: array
    type: object
...
  models.UpdateCountry:
    properties:
      code:
        type: string
      continent:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /country:
    get:
      consumes:
      - application/json
      description: Get List Country
      operationId: get_list_country
      parameters:
      - description: limit
        in: query
        name: limit
        type: number
      - description: offset
        in: query
        name: offset
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.GetListCounrtyResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Country
      tags:
      - Country
    post:
      consumes:
      - application/json
      description: Create Country
      operationId: create_country
      parameters:
      - description: CreateCountryRequestBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreateCountry'
      produces:
      - application/json
      responses:
        "200":
          description: CountryBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Country'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Country
      tags:
      - Country
    put:
      consumes:
      - application/json
      description: CountryUpdate
      operationId: country_update
      parameters:
      - description: UpdateCountryRequestBody
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCountry'
      produces:
      - application/json
      responses:
        "200":
          description: GetListCountryResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Country'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: CountryUpdate
      tags:
      - Country
  /country/{id}:
    delete:
      consumes:
      - application/json
      description: CountryDelete
      operationId: country_delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CountryPrimaryKey'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: CountryDelete
      tags:
      - Country
    get:
      consumes:
      - application/json
      description: Get By Id Country
      operationId: get_by_id_country
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetListCountrytResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Country'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By Id Country
      tags:
      - Country
  /upload:
    post:
      consumes:
      - application/json
      description: Upload file
      operationId: upload_file
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: table_slug
        in: query
        name: table_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: File
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Upload file
      tags:
      - File
swagger: "2.0"